+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.0.0
+ * Updated to `@animoca/ethereum-contracts-nft_staking:4.0.0`.
+ * Added a REVV escrowing based on staked weight for the staking.
+
 ## 0.4.0
  * Added `NFTRepairCentre`, a contract to manage defunct NFTs.
  * Updated to `@animoca/ethereum-contracts-core_library:3.1.1`.